gtest print 4

+++ app/Http/Controllers/PrintController.php
@@ -500,7 +500,6 @@ class PrintController extends Controller
                     ]);
                 }
 
-                // Ahora sí conviertes a base64
                 $archivosImpresion[] = base64_encode($content);
 
                 if (str_ends_with($archivo->nombre, '.zpl')) {
@@ -555,9 +554,18 @@ class PrintController extends Controller
                 $archivoFinal = $nombreArchivo;
             }
 
-            $modo = ($extension === 'zpl' || $marketplace->marketplace === 'MERCADOLIBRE') ? '-o raw' : '';
-            exec("lp -d {$ipImpresora} -n 1 {$modo} {$archivoFinal}");
-            $S = "lp -d {$ipImpresora} -n 1 {$modo} {$archivoFinal}";
+//            $modo = ($extension === 'zpl' || $marketplace->marketplace === 'MERCADOLIBRE') ? '-o raw' : '';
+//            exec("lp -d {$ipImpresora} -n 1 {$modo} {$archivoFinal}");
+            $fp = fsockopen($ipImpresora, 9100, $errno, $errstr, 10);
+            if (!$fp) {
+                return response()->json([
+                    'code' => 500,
+                    'message' => "Error al conectar a la impresora: $errstr ($errno)"
+                ]);
+            }
+
+            fwrite($fp, file_get_contents($archivoFinal));
+            fclose($fp);
             $outputs[] = $archivoFinal;
 
             if (file_exists($archivoFinal)) unlink($archivoFinal);
@@ -568,7 +576,6 @@ class PrintController extends Controller
             'code' => 200,
             'message' => 'Guías enviadas a impresión.',
             'outputs' => $outputs,
-            'a' => $S ??[]
         ]);
     }
